Create an empty TopicsService class test.
